rename tests package to benchmarks
